Big update
- Fixed RBM bugs
 -> Missing weights, bias etc. utilization
 -> Made gaussian-binary
- Restructured RBM algorithm
- Implemented scaling x_0 parameter
 -> Needs further iteration to be
- Implemented a plotting function for particle densities in 1, 2 and 3 dimensions for the purposes of analysis.
- Various other minor changes + some documentation
- In the current configuration, all methods don't yield very good results. Needs tuning of various parameters.

+++ Project/Code/energies.py
@@ -5,9 +5,9 @@ class Energies:
         Return the true energy of a two fermion system in two dimensions
         
         Returns:
-            int: 3
+            float: 3
         """
-        return 3
+        return 3.0
 
     def calogero_sutherland(N, omega, beta):
         """
